Fixing early season headaches + BOAT/WOAT
Did some fixing of the data handling that removes teams that haven't played any games which caused the sorting function to have a heart attack.  Doing this allowed me to remove some band-aid code in the sorting

There was also some redundant stuff going on with the Rank class and passing data to it.

Also did the rest of the BOAT and WOAT